ai-module/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
transformers
faiss-cpu
numpy
scikit-learn
python-multipart
requests